requirements.txt
gradio
openai-whisper
pytesseract
python-dotenv
Pillow
requests